documind-api-python/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
langchain>=0.3.25
langchain-core>=0.3.61
langchain-openai>=0.3.18
langchain-community>=0.3.24
langchain-text-splitters>=0.3.8
langgraph>=0.4.7
azure-storage-blob==12.25.1
azure-search-documents==11.6.0b9
tiktoken==0.9.0
pypdf==5.6.0
python-docx==1.1.2
python-dotenv==1.1.0
pydantic==2.11.3
python-multipart>=0.0.18
aiohttp>=3.9.0
